requirements.txt
certifi==2024.12.14
numpy==2.2.0
pandas==2.2.3
polygon-api-client==1.14.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
tzdata==2024.2
urllib3==2.2.3
websockets==12.0